chore(java): Upgrade IR dependency to v55 (#6092)
* chore(java): Upgrade IR dependency to v55

* Advance seed versions to v55 IR in java

* Cut a version for each java generator

* chore: update changelog

* Fix compilation

* remove comment

* Fix spring fixture

* Run model fixture

---------

+++ seed/java-model/object/src/main/java/com/seed/object/model/Type.java
@@ -61,11 +61,11 @@ public final class Type {
 
     private final Name nineteen;
 
-    private final int twenty;
+    private final long twenty;
 
     private final long twentyone;
 
-    private final double twentytwo;
+    private final float twentytwo;
 
     private final BigInteger twentythree;
 
@@ -88,9 +88,9 @@ public final class Type {
             List<Map<String, Integer>> sixteen,
             List<Optional<UUID>> seventeen,
             Name nineteen,
-            int twenty,
+            long twenty,
             long twentyone,
-            double twentytwo,
+            float twentytwo,
             BigInteger twentythree) {
         this.one = one;
         this.two = two;
@@ -212,7 +212,7 @@ public final class Type {
     }
 
     @JsonProperty("twenty")
-    public int getTwenty() {
+    public long getTwenty() {
         return twenty;
     }
 
@@ -222,7 +222,7 @@ public final class Type {
     }
 
     @JsonProperty("twentytwo")
-    public double getTwentytwo() {
+    public float getTwentytwo() {
         return twentytwo;
     }
 
@@ -345,7 +345,7 @@ public final class Type {
     }
 
     public interface TwentyStage {
-        TwentyoneStage twenty(int twenty);
+        TwentyoneStage twenty(long twenty);
     }
 
     public interface TwentyoneStage {
@@ -353,7 +353,7 @@ public final class Type {
     }
 
     public interface TwentytwoStage {
-        TwentythreeStage twentytwo(double twentytwo);
+        TwentythreeStage twentytwo(float twentytwo);
     }
 
     public interface TwentythreeStage {
@@ -444,11 +444,11 @@ public final class Type {
 
         private Name nineteen;
 
-        private int twenty;
+        private long twenty;
 
         private long twentyone;
 
-        private double twentytwo;
+        private float twentytwo;
 
         private BigInteger twentythree;
 
@@ -574,7 +574,7 @@ public final class Type {
 
         @java.lang.Override
         @JsonSetter("twenty")
-        public TwentyoneStage twenty(int twenty) {
+        public TwentyoneStage twenty(long twenty) {
             this.twenty = twenty;
             return this;
         }
@@ -588,7 +588,7 @@ public final class Type {
 
         @java.lang.Override
         @JsonSetter("twentytwo")
-        public TwentythreeStage twentytwo(double twentytwo) {
+        public TwentythreeStage twentytwo(float twentytwo) {
             this.twentytwo = twentytwo;
             return this;
         }
